Change sections to pages
The "About", "License", and "Privacy" pages were being rendered as
sections, even though they were actually pages.  This commit changes
them to pages, which allows simplifying the section template (because it
no longer has to check if a section has pages inside it—now all sections
have pages, because that is the only time it makes sense to be a section
rather than a page).